requirements.txt
matplotlib>=3.6.2
numpy>=1.23.5
Pillow>=9.3.0
tqdm>=4.64.1
wandb>=0.13.5

torch>=1.7.0
scikit-learn>=1.3.2
opencv-python>=4.8.1.78
scikit-image>=0.21.0
PyYAML>=6.0.1